+++ packages/suspense-website/src/examples/RecordUtils/createPendingRecord.ts
@@ -1,0 +1,6 @@
+import { createDeferred, createPendingRecord } from "suspense";
+
+const abortController = new AbortController();
+const deferred = createDeferred<number>();
+
+createPendingRecord(deferred, abortController);