+++ packages/suspense-website/src/examples/RecordUtils/createPendingRecordData.ts
@@ -1,0 +1,6 @@
+import { createDeferred, createPendingRecordData } from "suspense";
+
+const abortController = new AbortController();
+const deferred = createDeferred<number>();
+
+const data = createPendingRecordData(deferred, abortController);