+++ packages/suspense-website/src/examples/RecordUtils/createRejectedRecord.ts
@@ -1,0 +1,3 @@
+import { createRejectedRecord } from "suspense";
+
+createRejectedRecord(new Error("This is an example"));