+++ packages/suspense-website/src/examples/RecordUtils/updateRecordToRejected.ts
@@ -1,0 +1,9 @@
+import { createPendingRecord } from "suspense";
+
+const record = createPendingRecord();
+
+// REMOVE_BEFORE
+
+import { updateRecordToRejected } from "suspense";
+
+updateRecordToRejected(record, new Error("Something went wrong"));